requirements.txt
flask
elasticsearch
transformers
pandas
pytest
requests
flake8
black
logging
python-logging-loki

